+++ apps/meteor/app/wordpress/server/index.ts
@@ -1,1 +1,1 @@
-import '../lib/common';
+import './lib';
